chapiter 12

+++ Classes/Ijdb/Controllers/Joke.php
@@ -3,15 +3,18 @@
 namespace Ijdb\Controllers;
 
 use \Ninja\DatabaseTable;
+use Ninja\Routes;
+use Ninja\Authentication;
 
 class Joke
 {
     private $authorsTable;
     private $jokesTable;
-    public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable)
-    {
+    public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication)
+     {
         $this->jokesTable = $jokesTable;
         $this->authorsTable = $authorsTable;
+        $this->authentication = $authentication;
     }
 
     public function home()
@@ -24,38 +27,47 @@ class Joke
         $result = $this->jokesTable->findAll();
         $jokes = [];
         foreach ($result as $joke) {
-            $author =
-                $this->authorsTable->findById($joke['authorid']);
+            $author = $this->authorsTable->findById($joke['authorid']);
             $jokes[] = [
                 'id' => $joke['id'],
                 'joketext' => $joke['joketext'],
                 'jokedate' => $joke['jokedate'],
                 'name' => $author['name'],
-                'email' => $author['email']
+                'email' => $author['email'],
+                'authorid' => $author['id']
             ];
         }
         $title = 'Joke list';
         $totalJokes = $this->jokesTable->total();
+        $author = $this->authentication->getUser();
         return [
             'template' => 'jokes.html.php',
             'title' => $title,
             'variables' => [
                 'totalJokes' => $totalJokes,
-                'jokes' => $jokes
+                'jokes' => $jokes,
+                'userId' => $author['id'] ?? null
             ]
         ];
     }
     public function saveEdit()
     {
-        $joke['joketext'] = $_POST['joketext'];
+        $author = $this->authentication->getUser();
+        if (isset($_GET['id'])) {
+            $joke = $this->jokesTable->findById($_GET['id']);
+            if ($joke['authorid'] != $author['id']) {
+                return;
+            }
+        }
+        $joke = $_POST['joke'];
         $joke['jokedate'] = new \DateTime();
-        $joke['id'] = $_POST['jokeid'];
-        $joke['authorid'] = 1;
+        $joke['authorid'] = $author['id'];
         $this->jokesTable->save($joke);
         header('location: /joke/list');
     }
     public function edit()
     {
+        $author = $this->authentication->getUser();
         if (isset($_GET['id'])) {
             $joke = $this->jokesTable->findById($_GET['id']);
         }
@@ -63,11 +75,19 @@ class Joke
         return [
             'template' => 'editjoke.html.php',
             'title' => $title,
-            'variables' => ['joke' => $joke ?? null]
+            'variables' => [
+                'joke' => $joke ?? null,
+                'userId' => $author['id'] ?? null
+            ]
         ];
     }
     public function delete()
     {
+        $author = $this->authentication->getUser();
+        $joke = $this->jokesTable->findById($_POST['id']);
+        if ($joke['authorid'] != $author['id']) {
+            return;
+        }
         $this->jokesTable->delete($_POST['id']);
         header('location: /joke/list');
     }
